+++ spec/unit/formatters/error_inspectors/api_error_formatting_spec.rb
@@ -29,19 +29,21 @@ describe Chef::Formatters::APIErrorFormatting do
 
 
   context "when describe_406_error is called" do
-    context "when response.body['error'] == 'invalid-x-ops-server-api-version'" do
+    context "when response['x-ops-server-api-version'] exists" do
       let(:min_version) { "2" }
       let(:max_version) { "5" }
+      let(:request_version) { "30" }
       let(:return_hash) {
         {
-          "error" => "invalid-x-ops-server-api-version",
           "min_version" => min_version,
-          "max_version" => max_version
+          "max_version" => max_version,
+          "request_version" => request_version
         }
       }
 
       before do
-        allow(Chef::JSONCompat).to receive(:from_json).and_return(return_hash)
+        # mock out the header
+        allow(response).to receive(:[]).with('x-ops-server-api-version').and_return(Chef::JSONCompat.to_json(return_hash))
       end
 
       it "prints an error about client and server API version incompatibility with a min API version" do
@@ -53,17 +55,17 @@ describe Chef::Formatters::APIErrorFormatting do
         expect(error_description).to receive(:section).with("Incompatible server API version:",/a max API version of #{max_version}/)
         class_instance.describe_406_error(error_description, response)
       end
+
+      it "prints an error describing the request API version" do
+        expect(error_description).to receive(:section).with("Incompatible server API version:",/a request with an API version of #{request_version}/)
+        class_instance.describe_406_error(error_description, response)
+      end
     end
 
     context "when response.body['error'] != 'invalid-x-ops-server-api-version'" do
-      let(:return_hash) {
-        {
-          "error" => "some-other-error"
-        }
-      }
 
       before do
-        allow(Chef::JSONCompat).to receive(:from_json).and_return(return_hash)
+        allow(response).to receive(:[]).with('x-ops-server-api-version').and_return(nil)
       end
 
       it "forwards the error_description to describe_http_error" do